requirements.txt
aiohttp
datapizza-ai
datapizza-ai-clients-openai-like
python-dotenv
